Added basic test, coverage, github actions, improved readme. `mapengraver` is a work in progress title...
