+++ .gitignore
@@ -7,3 +7,5 @@
 ### Python ###
 __pycache__/
 venv/
+.pytest_cache
+.coverage
