+++ pyproject.toml
@@ -14,7 +14,18 @@ tqdm = "^4.58.0"
 flake8 = "^3.8.4"
 pytest = "^6.2.2"
 coverage = "^5.5"
+pytest-cov = "^2.11.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.coverage.paths]
+source = ["src", "*/site-packages"]
+
+[tool.coverage.run]
+branch = true
+source = ["mapengraver"]
+
+[tool.coverage.report]
+show_missing = true
