--- requirements.txt
@@ -1,6 +1,0 @@
-CairoSVG==2.3.1
-PyYAML==5.1
-Shapely==1.7a1
-cairocffi==1.0.2
-pyproj==2.1.3
-tqdm==4.31.1
