+++ tests/canvas/test_canvas.py
@@ -1,0 +1,8 @@
+import unittest
+
+from mapengraver.canvas import Canvas
+
+
+class TestCanvas(unittest.TestCase):
+    def test_init(self):
+        Canvas('png', 123, 134, 20)
